make service provider boosting module
It didn't feel quite correct to have half the service provider boosting
code in lib.rs and the other in location.rs.

The constants have to with location trust scores, but they do not get
used there.
